# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Document API Tiện Lợi Xanh",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "category"
                ],
                "summary": "Lấy danh sách danh mục",
                "description": "API này trả về toàn bộ danh sách các danh mục có trong hệ thống.",
                "operationId": "ba2f88513b5f6d9f04554acafeaaa2ac",
                "responses": {
                    "200": {
                        "description": "Lấy danh mục thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Danh mục A"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Mô tả của danh mục A"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lỗi định dạng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Lỗi định dạng."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/{categoryId}/products": {
            "get": {
                "tags": [
                    "category"
                ],
                "summary": "Lấy sản phẩm theo danh mục",
                "description": "API này lấy danh sách sản phẩm dựa trên ID của danh mục và chỉ trả về các sản phẩm và đơn vị có trạng thái active.",
                "operationId": "41064278c81d466a4acf56ac90aad2aa",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy sản phẩm theo danh mục thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "cate_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Sản phẩm A"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 100000
                                            },
                                            "sale": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 90000
                                            },
                                            "img": {
                                                "type": "string",
                                                "example": "url/image.jpg"
                                            },
                                            "quantity": {
                                                "type": "integer",
                                                "example": 50
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Mô tả sản phẩm A"
                                            },
                                            "made": {
                                                "type": "string",
                                                "example": "Vietnam"
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "category": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Danh mục A"
                                                    },
                                                    "units": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Đơn vị A"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Danh mục không tồn tại để tìm thấy sản phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Danh mục không tồn tại để tìm thấy sản phẩm!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Nhập token JWT của bạn trong trường này",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}